[Webview] Convert some straggling jint->jlong field for 64 bit.
Also roll frameworks/webview to include https://android-review.googlesource.com/#/c/88751/

BUG=354405

Review URL: https://codereview.chromium.org/205533006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258852 0039d316-1c4b-4281-b951-d872f2087c98

+++ android_webview/java/src/org/chromium/android_webview/AwSettings.java
@@ -257,7 +257,7 @@ public class AwSettings {
         }
     }
 
-    void setWebContents(int nativeWebContents) {
+    void setWebContents(long nativeWebContents) {
         synchronized (mAwSettingsLock) {
             if (mNativeAwSettings != 0) {
                 nativeDestroy(mNativeAwSettings);
